Move token format functions into their own file.
